villain: ExactHeatbath silently widened the integer field n to float64
L.zeros defaults to dtype=float, so the scratch 0-form z was float; d
faithfully preserves the dtype it is handed, making dn float, and
'n = n + dn' then promoted the INTEGER field n to float64.  A single sweep
was enough, and the Hammer runs one every step -- so everything downstream
treating n as exactly integral (h5 dtype, equality, mod-W arithmetic, the
no-intersection integer wedge) had been working on floats.

Nothing numerical changes: ksel is already an int from _draw, so only the
container was wrong.  Verified by forcing the old float path back and
comparing chains under a shared seed -- bit-identical over 40 sweeps.

Found by test_villain_hammer_preserves_field_dtypes, which was failing before
today's surface_worm work and is unrelated to it.  Swept all thirteen Villain
generators: ExactHeatbath was the only one with the pattern (the other
L.zeros(0) call sites build scratch forms for the float field phi, which is
correct).  'n = n + dn' rather than 'n += dn' is what let it through; an
in-place add would have raised instead of widening.

Two regression tests: one on the generator directly (so it still bites if
Hammer's composition changes), one asserting the fix is numerically a no-op.

+++ supervillain/generator/villain/exact_heatbath.py
@@ -132,11 +132,19 @@ class ExactHeatbath(ReadWriteable, Generator):
             kstar = g[0, *color] / (4 * np.pi * D)
             ksel = self._draw(kstar)
 
-            change_z = L.zeros(0)
+            # z must carry n's integer dtype: L.zeros defaults to float, and since d
+            # faithfully preserves the dtype it is handed, a float z makes dn float and
+            # silently widens the *integer* field n to float64 on `n = n + dn`.  The
+            # drawn values are integral either way (ksel comes from _draw as an int),
+            # so nothing numerical changes -- but everything downstream that treats n
+            # as exactly integral (h5 dtype, equality, mod-W arithmetic) would be
+            # working on floats after a single Hammer sweep.
+            change_z = L.zeros(0, dtype=n.dtype)
             change_z[0, *color] = ksel
 
             dn = d(change_z)
             n = n + dn
+            # r is the float residual dφ − 2πn, so this addition is float by design
             r = r - 2 * np.pi * dn
 
         return cfg | {'n': n}
